requirements.txt
open3d
torch
ruamel.yaml